rename, LF newlines, add CHANGELOG
