+++ CHANGELOG.md
@@ -1,0 +1,12 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a
+Changelog](https://keepachangelog.com/en/1.0.0/).
+
+## [Unreleased] - 2019-05-01
+(Poikilos)
+### Changed
+* renamed project
+* changed newlines to LF for linux
+* simplified README.md
